% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_cash}
\alias{set_cash}
\title{Title}
\usage{
set_cash(
  timesteps = timesteps,
  exp_dret = 0.01,
  exp_dsd = 0,
  from = 1,
  to = length(timesteps$timestep),
  amount = NULL,
  percentage = NULL
)
}
\arguments{
\item{percentage}{}
}
\value{
list
}
\description{
Title
}
