% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_expense}
\alias{set_expense}
\title{Set Expense}
\usage{
set_expense(timesteps, expense0, exp_inflation = 0, exp_inflsd = 0)
}
\arguments{
\item{inflation}{}
}
\value{
expense
}
\description{
Set Expense
}
