% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_income}
\alias{set_income}
\title{Title}
\usage{
set_income(timesteps, income0, exp_increase = 0, exp_increasesd = 0)
}
\arguments{
\item{exp_increasesd}{}
}
\value{
income list
}
\description{
Title
}
