% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_investment}
\alias{set_investment}
\title{Title}
\usage{
set_investment(
  name = NULL,
  timesteps = timesteps,
  exp_pret,
  exp_psd,
  exp_dret,
  exp_dsd,
  from = 1,
  to = length(timesteps$timestep),
  percentage = NULL,
  amount = NULL
)
}
\arguments{
\item{amount}{}
}
\value{
list
}
\description{
Title
}
