% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_timesteps}
\alias{set_timesteps}
\title{Set Timesteps}
\usage{
set_timesteps(
  start = ceiling_date(Sys.Date(), by),
  by = "year",
  n = 20,
  date_of_birth = NULL
)
}
\arguments{
\item{start}{data, default is ceilinged current date}

\item{by}{"year" (default) or "month"}

\item{n}{number of periods out}

\item{date_of_birth}{(optional) data of birth, used for computing age}
}
\value{
timesteps
}
\description{
Set Timesteps
}
