% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_wealth}
\alias{simulate_wealth}
\title{Title}
\usage{
simulate_wealth(
  timesteps,
  wealth0,
  expense,
  income,
  cash,
  correlations = NULL,
  investments,
  incometax,
  wealthtax
)
}
\arguments{
\item{wealthtax}{}
}
\value{
x
}
\description{
Title
}
